Cleaned up directories
