+++ lib/ECommerce/Views/UserView.php
@@ -4,18 +4,29 @@
 namespace ECommerce\Views;
 
 use ECommerce\Site;
+use ECommerce\Tables\Users;
 
 class UserView extends View {
 
-    public function __construct(Site $site) {
-        $this->setTitle("Pet Pack, LLC. User");
+    private $site;
+    private $users;
 
-        $this->addLink("users.php", "Users");
-        $this->addLink("staff.php", "Staff");
+    public function __construct(Site $site) {
+        $this->site = $site;
+        $this->users = new Users($this->site);
+
+        $this->setTitle("Users");
+
+        $this->addLink("./staff.php", "Staff");
         $this->addLink("post/logout.php", "Log Out");
     }
 
     public function present() {
+        echo $this->presentUsers();
+        echo $this->presentAddUser();
+    }
+
+    public function presentAddUser() {
         return <<<HTML
 <form>
 	<fieldset>
@@ -55,5 +66,46 @@ class UserView extends View {
 	</fieldset>
 </form>
 HTML;
+    }
+
+    public function presentUsers() {
+        $users = $this->users->getAll();
+        $html = <<<HTML
+<form class="table">
+	<p>
+	<input type="submit" name="add" id="add" value="Add">
+	<input type="submit" name="edit" id="edit" value="Edit">
+	<input type="submit" name="delete" id="delete" value="Delete">
+	</p>
+	<table>
+		<tr>
+			<th>&nbsp;</th>
+			<th>Name</th>
+			<th>Email</th>
+			<th>Role</th>
+		</tr>
+HTML;
+
+        foreach ($users as $user) {
+            $name = $user["name"];
+            $email = $user["email"];
+            $role = $user["role"];
+            $html .= <<<HTML
+		<tr>
+			<td><input type="radio" name="user"></td>
+			<td>$name</td>
+			<td>$email</td>
+			<td>$role</td>
+		</tr>
+HTML;
+
+        }
+
+        $html .= <<<HTML
+	</table>
+</form>
+HTML;
+        return $html;
+
     }
 }
