+++ post/add-product.php
@@ -1,5 +1,5 @@
 <?php
 require '../lib/site.inc.php';
 
-$controller = new ECommerce\Controllers\AddProductController($site, $_POST);
+$controller = new ECommerce\Controllers\ProductController($site, $_POST);
 echo $controller->getResult();
