+++ post/add-sub-cat.php
@@ -1,4 +1,4 @@
 <?php
 require '../lib/site.inc.php';
-$controller = new ECommerce\Controllers\AddSubCatController($site, $_POST, $_FILES);
+$controller = new ECommerce\Controllers\SubCatController($site, $_POST, $_FILES);
 echo $controller->getResult();
