+++ post/add-top-cat.php
@@ -1,5 +1,5 @@
 <?php
 require '../lib/site.inc.php';
 
-$controller = new ECommerce\Controllers\AddTopCatController($site, $_POST);
+$controller = new ECommerce\Controllers\TopCatController($site, $_POST);
 echo $controller->getResult();
