+++ src/js/Product.js
@@ -1,13 +1,13 @@
 import $ from 'jquery';
 import {parse_json} from "./parse_json";
 
-export const AddProduct = function () {
+export const Product = function () {
 
     $("form#add-product").submit(function (event) {
         event.preventDefault();
 
         $.ajax({
-            url: "post/add-product.php",
+            url: "post/product.php",
             data: $(this).serialize(),
             method: "POST",
             success: function(data) {
