+++ src/js/index.js
@@ -1,17 +1,15 @@
 import $ from 'jquery';
 
 import {Login} from "./Login";
-import {Users} from "./Users";
-import {AddTopCat} from "./AddTopCat";
-import {AddSubCat} from "./AddSubCat";
-import {AddProduct} from "./AddProduct";
+import {TopCat} from "./TopCat";
+import {SubCat} from "./SubCat";
+import {Product} from "./Product";
 
 import '../scss/Home.scss';
 
 $(document).ready(function () {
    new Login();
-   new Users();
-   new AddTopCat();
-   new AddSubCat();
-   new AddProduct();
+   new TopCat();
+   new SubCat();
+   new Product();
 });
